pyproject.toml
[project]
name = "data-intake-service"
version = "0.1.0"
description = "Data Intake Service"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy>=2.0,<3",
    "alembic>=1.14,<2",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "httpx>=0.28,<1",
    "python-json-logger>=2.0,<3",
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3,<10",
    "pytest-asyncio>=1.0,<2",
    "pytest-cov>=6,<8",
    "ruff>=0.8,<1",
    "mypy>=1.0,<2",
    "pip-audit>=2.7,<3",
    "pre-commit>=4.0,<5",
    "commitizen>=4.0,<5",
    "types-PyYAML>=6.0,<7",
]

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "alembic*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["--strict-markers", "--strict-config", "-ra"]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/__init__.py", "*/alembic/*"]
branch = true
relative_files = true

[tool.coverage.report]
precision = 2
show_missing = true
fail_under = 70.0
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI uses this pattern)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG", "S101"]  # Allow unused arguments and asserts in tests
"alembic/versions/*.py" = ["ALL"]  # Skip linting migration files
"alembic/env.py" = ["F401"]  # Allow unused imports for model registration

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = ["alembic.*", "sqlalchemy.*", "yaml"]
ignore_missing_imports = true
